Changed carousel button from click to mousedown and touchstart
This is a fix to make swiping appear to work when the swipe start occurs within
the buttons. It also makes it feel more responsive.

This does introduce a bug when using the firefox developer tools, and when touch
emulation is enabled it fires both a mouse and a touch event. A possible fix
would be to lock the button presses during the snap animation, but that reduces
percieved responsiveness.